fix (setup.sh): Use a base URL of http:// rather than https://
The docker-compose setup exposes port 80 on a random host port, however the base URL is set to https (and thus 443). This renders the site inaccessible by default.

This commit changes the base URL to http, so that the IP of the container can be added to the hostfile and the site accessed as normal.